+++ src/index.ts
@@ -1,13 +1,27 @@
 export { Client } from "./client";
-export { Environment } from "./environment";
+export { Environment, type EnvironmentType } from "./environment";
 export * from "./exception";
-export {
+export type {
   CardRequest,
   CardResponse,
   CheckResponse,
+  Currency,
   MobileRequest,
   MobileResponse,
   PayoutRequest,
   PayoutResponse,
   Transaction,
+} from "./schemas";
+export {
+  CardRequestSchema,
+  CardResponseSchema,
+  CheckResponseSchema,
+  CurrencySchema,
+  MobileRequestSchema,
+  MobileResponseSchema,
+  PayoutRequestSchema,
+  PayoutResponseSchema,
+  Status,
+  StatusSchema,
+  TransactionSchema,
 } from "./schemas";